Took out a room template which each room calls and passes in specific images and texts for that room. Improved grid to show full room pictures and onHover effects plus nicer spacing for smaller screens
